requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
aiologger==0.6.1
aiomysql==0.0.22
APScheduler==3.9.1
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
discord.py==1.7.3
future==0.18.2
greenlet==1.1.2
idna==3.3
marshmallow==3.15.0
multidict==6.0.2
packaging==21.3
PyMySQL==0.9.3
pyparsing==3.0.7
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
treelib==1.6.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
yarl==1.7.2
